+++ guidedmodules/urls.py
@@ -6,6 +6,15 @@ admin.autodiscover()
 import guidedmodules.views
 
 urlpatterns = [
+
+    # multi-question authoring tool
+    url(r'^module/(\d+)/questions$', guidedmodules.views.show_module_questions, name="show_module_questions"),
+    url(r'^_authoring_tool/edit-question2$', guidedmodules.views.authoring_edit_question2),
+    url(r'^module/(\d+)/artifact/new$', guidedmodules.views.add_module_artifact, name="add_module_artifact"),
+    url(r'^module/(\d+)/artifact/([\w_-]+)$', guidedmodules.views.show_module_artifact, name="show_module_artifact"),
+    url(r'^_authoring_tool/edit-artifact$', guidedmodules.views.authoring_edit_artifact, name="authoring_edit_artifact"),
+
+    # question management urls
     url(r'^(\d+)/([\w_-]+)(/_save)()$', guidedmodules.views.save_answer),
     url(r'^(\d+)/([\w_-]+)(/question/)([\w_-]+)$', guidedmodules.views.show_question),
     url(r'^(\d+)/([\w_-]+)(/question/)([\w_-]+)/history/(\d+)/media$', guidedmodules.views.download_answer_file),
@@ -20,7 +29,7 @@ urlpatterns = [
     url(r'^_start_discussion', guidedmodules.views.start_a_discussion, name="start_a_discussion"),
     url(r'^analytics$', guidedmodules.views.analytics, name="guidedmodules_analytics"),
     url(r'^_authoring_tool/new-question$', guidedmodules.views.authoring_new_question),
-    url(r'^_authoring_tool/edit-question$', guidedmodules.views.authoring_edit_question),
+    url(r'^_authoring_tool/new-question2$', guidedmodules.views.authoring_new_question2),
     url(r'^_authoring_tool/edit-module$', guidedmodules.views.authoring_edit_module),
     url(r'^_authoring_tool/create-app-project$', guidedmodules.views.authoring_create_q),
     url(r'^_authoring_tool/import-appsource$', guidedmodules.views.authoring_import_appsource),