+++ siteapp/static/js/authoring_tool.js
@@ -1,7 +1,6 @@
 function init_authoring_tool(state) {
   // Save state. Should be:
   // {
-  //  task: task_id,
   //  module: { module spec data },
   //  questions: {
   //    question_key: {
@@ -94,13 +93,15 @@ function init_authoring_tool_autocomplete(elem, expr_type) {
   });
 }
 
-function show_question_authoring_tool(question_key) {
+function show_question_authoring_tool(question_id, question_key) {
   // Initialize form state.
   window.q_authoring_tool_state.current_question = question_key;
   var qinfo = window.q_authoring_tool_state.questions[question_key];
   var spec = JSON.parse(JSON.stringify(qinfo.spec)); // clone since we delete properties
+
+  $('#at_q_id').val(question_id);
   $('#authoring_tool_qid').val(question_key); delete spec.id;
-  $('#authoring_tool_qtitle').val(spec.title); delete spec.title;
+  $('#authoring_tool_q_title').val(spec.title); delete spec.title;
   $('#authoring_tool_qprompt').val(spec.prompt); delete spec.prompt;
   $('#authoring_tool_type').val(spec.type)
     .change(); // trigger event that hides/shows the appropriate fields
@@ -136,7 +137,6 @@ function show_question_authoring_tool(question_key) {
   if (Object.keys(spec).length == 0) spec = "";
   else spec = jsyaml.safeDump(spec);
   $('#authoring_tool_qspec').val(spec);
-
   // Show modal.
   $('#question_authoring_tool').modal();
 }
@@ -215,10 +215,12 @@ function authoring_tool_download_app_project(task_id, is_project_page) {
 
 function authoring_tool_save_question() {
   var data = $('#question_authoring_tool form').serializeArray();
-  data.push( { name: "task", value: q_authoring_tool_state.task } );
+  // data.push( { name: "task", value: q_authoring_tool_state.task } );
   data.push( { name: "question", value: q_authoring_tool_state.current_question } );
+  console.log('authoring_tool_save_question')
+  console.log(data)
   ajax_with_indicator({
-      url: "/tasks/_authoring_tool/edit-question",
+      url: "/tasks/_authoring_tool/edit-question2",
       method: "POST",
       data: data,
       keep_indicator_forever: true, // keep the ajax indicator up forever --- it'll go away when we issue the redirect
@@ -267,6 +269,18 @@ function authoring_tool_new_question(task_id, question_id, is_project_page) {
   })
 }
 
+function authoring_tool_new_question2(question_id) {
+  ajax_with_indicator({
+      url: "/tasks/_authoring_tool/new-question2",
+      method: "POST",
+      data: { "question_id": question_id },
+      keep_indicator_forever: true, // keep the ajax indicator up forever --- it'll go away when we issue the redirect
+      success: function(res) {
+        window.location = res.redirect;
+      }
+  })
+}
+
 function show_authoring_tool_module_editor() {
   // Initialize form state for the compliance app catalog information
   // field and the module specification field.
@@ -296,7 +310,28 @@ function authoring_tool_save_module() {
 }
 
 function authoring_tool_create_q_form(argument) {
+  console.log(argument);
   $('#create_q_authoring_tool').modal();
+  console.log(Object.keys(argument))
+  if (Object.keys(argument).includes('app_id')) {
+    $('#authoring_tool_app_id').val(argument['app_id']);
+  }
+  if (Object.keys(argument).includes('app_title')) {
+    console.log(argument);
+    $('#authoring_tool_q_2_title').val(argument['app_title']+" Copy");
+    var slug = $('#authoring_tool_q_2_title').val();
+    slug = slug.toLowerCase().replace(/[^a-z0-9--]+/g, "_");
+    $('#authoring_tool_q_name').val(slug);
+  }
+  if (Object.keys(argument).includes('app_description')) {
+    $('#authoring_tool_q_description').val(argument['app_description'].replace('<p>','').replace('</p>',''));
+  }
+  if (Object.keys(argument).includes('app_category_title')) {
+    $('#authoring_tool_q_category').val(argument['app_category_title']);
+  }
+  if (Object.keys(argument).includes('modal_title')) {
+    $('#create_q_authoring_tool_title').html(argument['modal_title']);
+  }
 }
 
 function authoring_tool_create_q() {