+++ CHANGELOG.md
@@ -4,12 +4,18 @@ GovReady-Q Release Notes
 v0.9.11.7-dev (November xx, 2021)
 --------------------------------
 
+**Feature changes**
+
+* Enabling batch viewing of questions for easier questionnaire authoring.
+* Enabling cloning entire templates in template library.
+
 **Developer changes**
 
 * Add Django `nlp` app to system to support Natural Language Processing of SSPs and statements.
 * Include spaCy libraries as part of build.
 * Include initial, simplified candidate entity recognition script.
 * Remove full text search of statements from component library search because it was slow and returned to many results.
+* Add serializers for Modules and ModuleQuestions
 
 **UI changes**
 
@@ -17,6 +23,7 @@ v0.9.11.7-dev (November xx, 2021)
 * Improve toast message appearance by limiting width.
 * Improve styling of project page rollovers make module actions clearer
 * Improve styling of template library. Use bootstrap panels for each item.
+* Remove authoring tool modal from task question page.
 
 
 v0.9.11.6 (October 13, 2021)